requirements.txt
asgiref==3.5.2
attrs==22.1.0
boto3
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.50.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
